Added sqlite storage logic

+++ internal/models/app.go
@@ -3,11 +3,11 @@ package models
 import "context"
 
 type App struct {
-	ID     int
+	ID     int64
 	Name   string
 	Secret string
 }
 
 type AppProvider interface {
-	App(ctx context.Context, id int) (App, error)
+	App(ctx context.Context, id int64) (App, error)
 }
